extract virtule-module into a independent util
