+++ packages/react-pages/src/node/utils/virtual-module/README.md
@@ -1,0 +1,4 @@
+# virtule modules
+
+This is an independent library to build, update and watch a virtule module graph.
+We can extract it into a npm package in the future.
