+++ packages/react-pages/src/node/utils/virtual-module/index.ts
@@ -1,0 +1,6 @@
+export { extractStaticData, File } from './utils'
+
+export { VirtualModuleGraph } from './VirtualModules'
+export type { ModuleListener, VirtuleModuleAPIs } from './VirtualModules'
+export { VirtualModulesManager } from './VirtualModulesManager'
+export type { FileHandlerAPIs } from './VirtualModulesManager'
