+++ packages/react-pages/src/node/utils/virtual-module/utils/extractStaticData.ts
@@ -1,6 +1,6 @@
 import { extract, parse } from 'jest-docblock'
 import grayMatter from 'gray-matter'
-import type { File } from './VirtualModulesManager'
+import { File } from './File'
 
 export async function extractStaticData(
   file: File
@@ -24,37 +24,6 @@ export async function extractStaticData(
       throw new Error(`unexpected extension name "${file.extname}"`)
   }
 }
-
-/**
- * track how many works are pending
- * to avoid returning half-finished page data
- */
-export class PendingList {
-  private pendingCount = 0
-  private subscribers: Array<() => void> = []
-
-  addPending(p: Promise<void>) {
-    this.pendingCount++
-    p.finally(() => {
-      this.pendingCount--
-      if (this.pendingCount === 0) {
-        // all pending works are finished
-        this.subscribers.forEach((notify) => notify())
-        this.subscribers.length = 0
-      }
-    })
-  }
-
-  subscribe(): Promise<void> {
-    if (this.pendingCount === 0) return Promise.resolve()
-    return new Promise((res) => {
-      this.subscribers.push(() => {
-        res()
-      })
-    })
-  }
-}
-
 function extractMarkdownTitle(code: string) {
   const match = code.match(/^# (.*)$/m)
   return match?.[1]
