+++ packages/react-pages/src/node/utils/virtual-module/utils/index.ts
@@ -1,0 +1,3 @@
+export { extractStaticData } from './extractStaticData'
+export { PendingState, PendingTaskCounter } from './PendingTaskCounter'
+export { File } from './File'
